+++ src/main/java/org/example/repository/AstartesMeleeWeaponRepository.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface AstartesMeleeWeaponRepository {
 
 
-    AstartesMeleeWeaponEntity get(Long id);
+    AstartesMeleeWeaponEntity get(int id);
     void add(AstartesMeleeWeaponEntity astartesMeleeWeaponEntity);
     void update(AstartesMeleeWeaponEntity astartesMeleeWeaponEntity);
     void remove(AstartesMeleeWeaponEntity astartesMeleeWeaponEntity);
